move optimizer code to scalar_fn_pushdown
