+++ vortex-duckdb/cpp/scalar_fn_pushdown.cpp
@@ -1,11 +1,9 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
-#include "optimizer.hpp"
-#include "table_function.hpp"
 #include "duckdb/catalog/catalog.hpp"
-#include "duckdb/main/config.hpp"
-#include "duckdb/main/capi/capi_internal.hpp"
 #include "duckdb/planner/operator/logical_projection.hpp"
+#include "scalar_fn_pushdown.hpp"
+#include "table_function.hpp"
 #include <optional>
 
 /**
@@ -14,24 +12,6 @@
  * If there are any functions left, this means they were not pushed down and
  * may produce conflicts (e.g. WHERE prefix("str", 'h')).
  */
-
-extern "C" duckdb_state duckdb_vx_optimizer_extension_register(duckdb_database ffi_db) {
-    D_ASSERT(ffi_db);
-    const DatabaseWrapper &wrapper = *reinterpret_cast<DatabaseWrapper *>(ffi_db);
-    DatabaseInstance &db = *wrapper.database->instance;
-    try {
-        DBConfig::GetConfig(db).GetCallbackManager().Register(VortexOptimizerExtension());
-    } catch (const std::exception &e) {
-        ErrorData data(e);
-        DUCKDB_LOG_ERROR(db, "Failed to create Vortex optimizer extension:\t" + data.Message());
-        return DuckDBError;
-    }
-    return DuckDBSuccess;
-}
-
-void VortexOptimizeFunction(OptimizerExtensionInput &input, LogicalOperatorPtr &plan) {
-    plan = TryPushdownScalarFunctions(input.context, std::move(plan));
-}
 
 LogicalOperatorPtr TryPushdownScalarFunctions(ClientContext &context, LogicalOperatorPtr plan) {
     Analyses analyses;
@@ -48,8 +28,7 @@ LogicalOperatorPtr TryPushdownScalarFunctions(ClientContext &context, LogicalOpe
             if (expr == nullptr) { // Conflict for column
                 continue;
             }
-            const TableColumnStorageIndex storage_index =
-                analysis.get.GetColumnIds()[column_index].GetPrimaryIndex();
+            const TableColumnStorageIndex storage_index = analysis.StorageIndex(column_index);
             TableFunctionProjectionExpressionInput input {analysis.get, *expr, storage_index};
             if (projection_expression_pushdown(context, input)) {
                 analysis.get.types[column_index] = expr->return_type;
@@ -247,4 +226,8 @@ ScalarFnCollect::ScalarFnCollect(Analyses &analyses, const Projections &projecti
 
 ScalarFnReplace::ScalarFnReplace(Analyses &analyses, const Projections &projections)
     : analyses(analyses), projections(projections) {
+}
+
+TableColumnStorageIndex GetAnalysis::StorageIndex(TableColumnScanIndex idx) const {
+    return get.GetColumnIds()[idx].GetPrimaryIndex();
 }
